+++ include/lbann/data_readers/data_reader_cifar10.hpp
@@ -65,8 +65,8 @@ public:
 
 protected:
   void set_defaults() override;
-  bool fetch_datum(CPUMat& X, int data_id, int mb_idx) override;
-  bool fetch_label(CPUMat& Y, int data_id, int mb_idx) override;
+  bool fetch_datum(CPUMat& X, uint64_t data_id, uint64_t mb_idx) override;
+  bool fetch_label(CPUMat& Y, uint64_t data_id, uint64_t mb_idx) override;
 
 private:
   /**
